app1/requirements.txt
matplotlib==3.5.1
pickle-mixin
pandas==1.4.2
scipy
Pillow==9.0.1
scikit-learn==1.0.2
openpyxl==3.0.9
Django
sahi==0.11.13
--extra-index-url https://download.pytorch.org/whl/cu117
torch==1.13.1+cu117
torchvision==0.14.1+cu117
torchaudio==0.13.1
yolov5==7.0.4
